+++ code/services/osservice.go
@@ -12,8 +12,8 @@ import (
     // "golang.org/x/sys/unix"
     "golang.org/x/sys/windows"
 
-	"github.com/Voltaic314/Data_Migration_Tool/responsehandler"
-	"github.com/Voltaic314/Data_Migration_Tool/filesystem"
+	"github.com/Voltaic314/Data_Migration_Tool/code/responsehandler"
+	"github.com/Voltaic314/Data_Migration_Tool/code/filesystem"
 )
 
 // OSService handles file and folder operations for the OS.
